+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix a test failing on macOS 11.3
 - Fix generation of inline:auto annotations in files with other inline annotations.
 - Fixes modifier access for things like `isLazy`, `isOptional`, `isConvienceInitializer`, `isFinal`
+- Fixes `isMutable` on subscripts
 
 ---
 ## 1.4.1